# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "lumino 记账本服务",
        "title": "Lumino",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/api/v1/category": {
            "get": {
                "tags": [
                    "图标"
                ],
                "summary": "获取图标",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID 图标id",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "用户id",
                        "name": "user_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "图标结果",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Category"
                            }
                        }
                    },
                    "400": {
                        "description": "请求体异常",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "服务端异常",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "图标"
                ],
                "summary": "修改图标",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID 图标id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "分类名称的中文示意",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "用户id",
                        "name": "user_id",
                        "in": "query"
                    },
                    {
                        "type": "file",
                        "description": "分类图标文件",
                        "name": "icon_file",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "请求体异常",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "服务端异常",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "图标"
                ],
                "summary": "注册图标",
                "parameters": [
                    {
                        "type": "string",
                        "description": "分类名称的中文示意",
                        "name": "name",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "用户id",
                        "name": "user_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "分类图标文件",
                        "name": "icon_file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "请求体异常",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "服务端异常",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "图标"
                ],
                "summary": "删除图标",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID 图标id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "请求体异常",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "服务端异常",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/friend": {
            "delete": {
                "tags": [
                    "朋友"
                ],
                "summary": "删除朋友",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "被邀请人",
                        "name": "invitee",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "邀请人",
                        "name": "inviter",
                        "in": "query"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "请求体异常",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "服务端异常",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/friend/invite": {
            "post": {
                "tags": [
                    "朋友"
                ],
                "summary": "邀请朋友",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "被邀请人",
                        "name": "invitee",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "邀请人",
                        "name": "inviter",
                        "in": "query"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "请求体异常",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "服务端异常",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.Category": {
            "type": "object",
            "properties": {
                "created_at": {
                    "description": "创建时间",
                    "type": "string"
                },
                "deleted_at": {
                    "description": "DeletedAt 删除时间",
                    "type": "string",
                    "example": "2025-03-26T00:00:00Z"
                },
                "icon_url": {
                    "description": "类别图标的本地地址",
                    "type": "string"
                },
                "id": {
                    "description": "主键id",
                    "type": "integer"
                },
                "name": {
                    "description": "分类名称的中文示意",
                    "type": "string"
                },
                "updated_at": {
                    "description": "更新时间",
                    "type": "string"
                },
                "user_id": {
                    "description": "用户id",
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}